requirements.txt
javalang
pycparser
esprima
xmltodict
anthropic
langchain_core
langchain_community
langchain_openai
langchain_anthropic
pydantic
langchain-google-genai
PyGithub
google-generativeai
httpx